moved elements to main library
